Add tpp.identity and tpp.matmul
Each tpp operation carries a trait composed of indexing_maps and iterator
types. The idea would be to use this information to pattern match the access
pattern and iterator a linalg.generic. Still unclear how to match the body
(e.g., the operation within the region of the linalg.generic).