Fix handling of `ldo` computation in tpp.identity -> xsmm conversion.
Previously, the stride was not properly taken into consideration.
This revision fixes this for the special case of the output for the tpp.identity conversion.

It is likely many other places need to be revisited on this model:
  1. check shape, type and rank
  2. extract strides from the memref type (or using memref.extract_strided_metadata)
  3. perform some checks on the expected most-minor stride.

This fix is enough to allow padded 2d convolutions to run through IREE.