Create TPP-RUN program
This program is a replacement for mlir-cpu-runner for a small number of
cases, where we want to quickly call a MLIR method with a wrapper
function that initializes the arguments and prints the output.

In addition, on benchmark mode, we also want to time the execution of
the kernel (but not the tensor setup, JIT compile times, etc).

For everything else, you should use mlir-cpu-runner and equivalent.

Tests were amended to use the right tool using the right command line
options for tpp-run as well as update options for tpp-opt.

For now, the features available are:
 * Reads MLIR file in `func` dialect (not `LLVM` dialect)
 * Detects / uses entry point passed as argument (or single function)
 * Creates a wrapper function that scans arguments and return values
 * Initializes arguments, prepare return values and calls the function
 * Prints the return value (whether it was returned or inplace)

What's still missing:
 * Expecting IR to be on a particular shape (memref) isn't very robust
 * Initializing arguments with splat(1.0), need to be random
 * Needs random seed to allow for predictable / comparable runs
 * Has no timer functionality yet
 * Doesn't try to figure out libxsmm/tpp library paths automaticaly

Bugs:
 * Inplace returns are still printing the original (arg) value
 * Can't update cmd-line args, needs hack, can't hav additional args

Future changes that may need MLIR (JitRunner) changes:
 * We need to access the cmd-line options
 * We need to time just the kernel execution, not setup, not JIT