Make `map_linalg_to_tpp` composable
Make `map_linalg_to_tpp` composable by avoiding invalidating the
handles. We don't require the op to be isolated from above anymore, as
we don't blindly apply rewrite patterns anymore. Add an optional filter
that allows controlling the subset of operations returned.

To make the op composable, expose a FailureOr<> for mapping linalg to
tpp. The mapping should never fail, but we need to avoid while(1) in the
rewriter; we need to revisit this function once we have a transform-only
flow.

Address #149