Factoring MLIRBench from main file
Changes how we lower MLIR to a set of helper functions. The API isn't
settled yet, because we still need to lower benchmark calls and
counters, which can complicate a bit, so it looks a bit of a mix between
a driver and a static helper class. Once we finish the functionality, we
can re-factor to a nices API, since it's small anyway, should be easy.

Needs https://reviews.llvm.org/D137861
Commit 1b99e8ba487b