Factoring MLIRBench from main file (#164)
Changes how we lower MLIR to a set of helper functions. The API isn't
settled yet, because we still need to lower benchmark calls and
counters, which can complicate a bit, so it looks a bit of a mix between
a driver and a static helper class. Once we finish the functionality, we
can re-factor to a nicer API. Since it's small, should be easy.

Working towards #74.