Benchmark kernels with simple statistics (#174)
This commit enables benchmarking of generic kernels (with variable 2D
tensor input variables and return vaule) by running it multiple times
and using a runtime library to time, accumulate and calculate some basic
mean+-dev statistics if the `-n` argument is greater than 1.

A few considerations:
* The runtime is an implementation of what we expect will be the `perf`
dialect. Once we have such a dialect, we should remove the perf logic
from the benchmark and just use that dialect.
* The restriction of 2D inputs are due to the printing via vector
extract. It should be reasonably easy to extend it to xD memrefs later.
* The API is still a bit clumsy. I don't want to change too many things
at the same time. So, once we consolidate on a working mechanism (`perf`
and `check` dialects, xD memrefs, etc), we can revisit the API.

Fixes #74

@hfp, this is the starting point of a generic benchmark runner for MLIR.