Change how loggers are created in benchmark harness
Before, all log lines were from the same place ("__main__") because they
were being passed as an object. A better approach is to get a new logger
for each class so it's clear where the logs are coming from. But that
also means we need to pass the loglevel to all classes, instead of the
logger itself.

Fixes #198