Perf dialect (#100)
Adds a new infrastructure dialect that facilitates code performance
benchmarking directly at the IR level.

The perf dialect introduces a set of operations that allows to collect
and process basic code runtime statistics.
The dialect positions itself as a high-level generic utility which
relies on support of a runtime that supplies platform dependent
functionality such as timers.