`tpp-run` -print not default (NFC)
The print functionality isn't good enough to be enabled by default and
may soon be replaced by checking output inside IR as a more stable
validation mechanism.

This commit doesn't change the functionality, but makes the option
disabled by default and adds it to the tests that rely on it.

This should make it easier for people to experiment with `tpp-run` until
it becomes more robust.