Benchmark using CMake, C++ classes
This refactors the benchmarks to simplify adding new code, commoning up
most logic in simple classes like Bench, Tensor and Config.

It also simplifies the build and execution of the benchmarks by building
the C++ benchmarks via CMake and only calling the binaries via the
driver.

There's a ninja benchmark target that calls the script in the right way,
depending on the three benchmarks, so you can just call `ninja
benchmark` and it will rebuild the C++ benchmarks and the MLIR tools (if
necessary) and run the benchmark correctly.

Fixes #199
Fixes #251