Add libraries to tpp-run to avoid --shared-libs arg
This facilitates the execution of our tests with tpp-run, that we can't
do with mlir-cpu-runner, by making it simpler to execute MLIR files
without havin to pass a build-time option that has the location of the
MLIR and TPP libraries.

This reduces the need for extra complexity in the harness.

Works towards #256