Partially retire linalgx pack and unpack (#238)
Partially retire linalgx.pack and linalgx.unpack, and switched to
upstreamed versions. Disabled test: `perf-to-loops.mlir`: rebasing to
the new LLVM version surface an issue tracked in #277

All other tests have been updated to use the upstream version (tensor.pack/tensor.unpack). 
For some transform tests, we check the IR at the tensor level instead of memref
as currently we don't have a way in transform to generalize tensor.pack
and tensor.unpack, I can follow up if necessary with some small
additions. 

The old operations linalgx.pack and linalgx.unpack are left
as some tests depend on them, but the ops are now restricted to memref
only and not used in the code-gen flow.