Introduce a pass to rewrite scf.for to scf.parallel
`scf.parallel` does not work at the tensor level. To maintain
parallelism at the tensor level, we have the following choices:
`scf.foreach_theread` or gml parallel loop in mlir-hlo. Preliminary
discussion seems that we don't like either options:
`scf.foreach_theread` does not have a lowering, and it is too 'thread'
like, and it is unclear how to lower it to OMP, while gml parallel loop requires
adding mlir-hlo as a submodule. If we exclude these two options, we
should either extend upstream `scf.parallel` on tensor or raise back
`scf.for` to `scf.parallel` after bufferization. This PR implements the
latter option, which is *wrong*, but it is the path with less friction.

We only raise loop marked as 'parallel'; other passes like
tile-consumer-and-fuse-producers are responsible for injecting the
attribute when a loop with parallel semantics is materialized.