Rework tile-consumer-fuse-producers (#278)
Initially, we used `tileConsumerAndFuseProducerGreedilyUsingSCFForOp` to
tile and fuse element-wise operation with matmul-like ops. However, the
use of `tileConsumerAndFuseProducerGreedilyUsingSCFForOp` is
inconvenient as we cannot control it. Commit:

https://github.com/llvm/llvm-project/commit/9db7d4edd82ed5bb3867bbe6093ed2c63b18425a
in LLVM exposes internal building blocks; thus, we can now implement our
version of `tileConsumerAndFuseProducerGreedilyUsingSCFForOp` but
retaining complete control on what gets tiled or fused.

We also provide a default, where the two-outermost parallel loops are
tiled by 32 and fused. The PR works for simple element-wise op + matmul
producers. They approach we use is similar in spirit to what IREE does
(thanks Mahesh).