Random initialization of tensors
Wrote a simple class structure to initialize tensors in various ways,
including the existing `dense<1.0>`, some odd patterns and two random
splats.

The random init creates a `dense<0x...>` with random values with a
library that can be used by both `tpp-run` and the C++ benchmarks.

Using the same seed, we can make sure we generate the same tensors on
both sides.

Fixes #229