Fix mapping to tpp.matmul (#316)
`isChainOfUnaryOpsFrom` checks for chain of unary operations but it does
not take into account the op type (i.e., mul or add). The type is taken
into account in `getSingleOpOfType` but it just requires to have an add
or a mul in the block. If there are other operations in the block
`getSingleOpOfType` does not catch it. To fix the mapping restrict the
block to have only three operations: add, mul and yield.