Require full-tiles when blocking (#320)
Require tiling factors for convolutions to create full tiles. We did not
have any checks on tile sizes for convolutions; this means that the user
could have passed a tile factor that did not create a full tile, and
`tensor.pack` was complaining as it does not support full tiles unless
padding is set. We do not set padding (at the moment) when blocking
convolutions.