Forward output of `linalg.fill` to `tensor.unpack` (#323)
Let's consider the following IR:

```
%e = tensor.empty() : tensor<2x2xf32>
%0 = linalg.fill ins(%cst : f32) outs(%e : tensor<2x2xf32>)
%1 = linalg.matmul ins( ... ) outs(%0 : tensor<2x2xf32>)
```

When blocking the matmul a `tensor.unpack` is inserted right after the
blocked matmul. The `tensor.unpack` will unpack the output of the matmul
to the orginal un-blocked output (%0) in this case. This is fine becuase
%0 will alias %e after bufferization but the IR looks better and some
canonicalization patterns are simplified if we unpack directly in %e.