Relax `SameTypeOperands` constraint for `tpp.add` (#297)
Requiring all operands to have the same type is too restrictive, and we
may lose mapping opportunities. For example, one operand can be a
strided memref like `memref<56x32xf32, strided<[32, 1], offset: ?>>`
while another, a non-strided one `memref<56x32xf32>`. The types of the
memref are different, but we still want to allow both in `tpp.add`.
Relax the operand check and now require each operand to have the same
*shape* and *strides*.