Add support for ReLU detection on dense / global (#335)
Resnet model uses a `constant dense<0.0>` at tensor level and `global
dense<0.0>` at memref level, so we need to support those two patterns.

Updated the Resnet bottleneck test to match ReLUs, as well as adding
more ReLU tests on its own tests and removing the Resnet Benchmark
because it was not quite right.

It took me a while to "get it right" on the Model side, so I think we
should just keep the models there for both IR and output checks
(contrary to my previous recommendation, so mea culpa).

Previous code was iterating through "previous users" which works for
some patterns but it's too lenient, especially on recursive patterns.

There are two remaining patterns that we don't catch because of that:
1. empty/alloc + fill(0.0) -> generic
2. same as above + non-writing ops -> generic

In the first case, the "previous user" works, as it detects `fill` as
the previous user, checks that it writes "all-zeros" and matches.

The second case is the adversarial generic case that breaks the
"previous user" assumption. For example:

```
%c0 = constant 0.0
%c1 = constant 1.0

// Fill + op
%0 = empty
%1 = fill(%c0) into %0
%2 = some_op(%1)

// Using %1 is ok in value but not ok in reference semantics
%2 = whatever... (this could change %1's memory)
%3 = linalg.generic ins(%1)

// Using %2 is only ok if `some_op` isn't changing the underlying value
%2 = copy(%1)
%2 = linalg.generic ins(%2) // OK
// oops
%2 = fill(%c1) into %1 // same as %0 really
%2 = linalg.generic ins(%2) // NOT OK!
```

But there are even more complicated cases that we'd match:
```
%c0 = constant 0.0
%c1 = constant 1.0

// Fill + op + op + generic
%0 = empty
%1 = fill(%c0) into %0
%2 = fill(%c1) into %0
%3 = copy(%1) // in reference semantics, this is already dense<1.0>
%4 = linalg.generic ins(%1)
%5 = some use of %2
```

In this last case, following "previous user" from `%4` would match
`copy(%1)` which has def as `fill(%c0)` which matches, but we ignore the
second fill, which cannot be a match in reference semantics.

If linalg-on-tensors were pure value semantics, we'd be ok. But it
isn't, so we're not ok.

This complication will be left for another patch, since we still don't
have a use case for it, but just laying it here so that we know why
those tests are still not matches ReLU.

Fixes #295