Allow to fuse from the last fusable consumer
Previously we were moving the anchor point of fusion up to the next
immediate fusbale consumer of a given conv or matmul-like operation.
This PR reworks the logic to shift the anchor point to the last fusbale
consumer, this avoid us to run pre-processing pass like linalg fusion on
tensor. In addition, requires to have at least to operations in the
worklist for tile and fuse to kick-in.