Splat replacement with random init
New option to `tpp-run` that allows us to replace splat tensors on the
kernel function with random init, so that we can write tests with
`dense<1.0>` and let the runner replace with actual random values (from
a normal distribution).

The option is `-splat-to-random` and needs `-seed` to be enabled, too.

Incidentally fixes #346