Relax consumer selection in tile and fuse (#355)
We used to require each consumer to have a single user, this is a bit
too restrictive, and we may lose fusion opportunities (i.e., the last
relu does not get fused as it has multiple users - return and print). So
change the candidate walk by going bottom up, not requiring a single
user. The bottom-up traversal allows us to insert candidates differently
and drop the user check; we know if an op is already part of a fusion
cluster if it is already visited.