Fix mapping for `tpp.identity`
The mapping logic for `tpp.identity` was too naive. So now we check the
affine maps and require them to be projected permutation. Additionally,
we ensure before mapping that the linalg op verifies "our" broadcasting
rules and rejects candidate ops that do not satisfy these requirements.
Current mapping is still not optimal, and more effort needs to be spent
on mapping and broadcast support.