Fix mapping to relu  (#370)
Add an extra operand (OperandInfo) to `isTppRelu` that returns the input
and output value. Drop `isUnaryOp` and `isBinaryOp` since this is not
always the case - a linalg.generic representing a relu can also be
ternary. The non-constant operand (the input of tpp relu) is obtained by
looking at the linalg region and extracting the OpOperand tied to the
block argument that represent a non constant value. We could follow a
similar approach for other tpps, only looking a the scalar operation
inside the region and get the input/output operands tied to the block
arguments.