Fix resnet50 bottleneck results return (#378)
Replaces copy-like linalg.generic with linalg.copy operation which does
not get removed by a combination of CSE on tensors and postprocessing.
This would lead to use-after-free problem as the result buffer gets
deallocated before return. The more explicit copy ensures that results
get correctly stored in the last function argument and can be printed
afterwards.