Adding BF16 MLP benchmarks
Adding some temporary, manually packed, BF16/VNNI benchmarks so we can
start tracking.

Because of bugs in MLIR bufferization, I had to disable LSAN on those
benchmakrs so they stop complaining about the temporary buffers leaked.
Also had to XFAIL those tests, as I don't know if it's possible to add
the environment to RUN lines (LLVM doens't seem to have this).

Of course this is bad, but we know what it is and how to fix.