Provide defaults for tile and fuse pass (#390)
* Fix `isTppMatmul` in case the operation is a linalg.matmul

* Rework `isBlockedMatmul`. The idea is to provide a better way to
understand if a given linalg.generic is a blocked matmul. Do do that we
check operands, affine maps and classify the loops. Each loops can
belong to three categories only I, J and K. The I loops are present in C
and A but not in B and are required to be parallel. The J loops are
present in C and B but not in A and must be parallels. The K loops are
present in A and B but not in C, and are reductions.

* Provide defaults for tile and fuse pass. linalg.matmul are blocked
with tile sizes of 32 along i and j. Blocked convolutions are tile and
fused along the three outermost loops to expose a BRGEMM. Similarly,
blocked matmul are tiled along the two outermost parallel loops.