Revert "Revert "Provide defaults for tile and fuse pass (#390)""
Re-apply commit 2db7058b0c6b8a9e218b240ec99f2b8a2a42b94d.

`isBlockedMatmul` assumed all the maps to be permutation maps, but
surprise, this is not the case for VNNI packing, where you have a
floor-div expression with constant on the B operand. Specially treat the
B operand by providing a walker that ensures each other dimension is an
AffineDimExpr, and there is one floor-div dimension with constant. This
dimension must be a reduction. The rest of the logic remains the same;
each loop must be classified in the I, J, or K bucket. I are parallels,
J is parallels, and K is reductions.