Re-work matchers
Second iteration on the matchers. The initial version needed to be more
flexible. To address these comments, the matchers expose only four
public predicates:

- Operation predicate using std::function<Operation*> callable objects.

- Input and output predicate (aka operand predicates) using
  std::function<OpOperand *, Operation*> callable objects. Note that the
operation pointer is needed to get operand information in Linalg.

- Dimension predicates using RangeDims and vector of expected iterators.
- Predicate on the region using std::function<bool(Operation *,
  SmallVector<Value> *captures)>

We gain flexibility using a set of composable predicates. Most of the
smartness is now done in the predicates and not within the matchers; the
user must provide 'sane' predicates, as only a few checks can now be
done at the matcher level.