XSMM document (#422)
This is a simplified XSMM dialect document. We're still hashing out the
details, but in theory, this is what we want to lower from TPP.

The flags aren't explained, because we're not quite sure how to work
them out, but they'll end up as an integer argument to the libxsmm calls
anyway.