Use matchers in linalg to tpp conversion
We still provided isTpp* methods, but the implementation now uses the
newly introduced structural matchers; this allows us to remove clutter
and have a nice syntax to define what we want to match. Matchers are
composed together (see `isTppOp` and `isTppBinary/Unary`).

We also update the operation traits for unary and binary.  All the unary
and binary tpp ops implement `BroadcastableShape` and
`UnitStrideInnerLoop`.