Do not allow scalar output in tpps
Tpp are memory-to-memory operations and we should not allow a scalar as
output memref. With this changes the tpp.relu and tpp.add verifier can
be removed as we check operand shape directly in TD. Update tests and
avoid checking operand prefix, this is automatically generate and there
is not need for us to check.