Parser/printer for perf ops
Adds a parser and printer for both bench and yield ops as parseTypeList
was crashing the parser with `grow_pod` on `emplace_back`. This is
likely an upstream problem that isn't being hit by upstream tests
because no one uses `Variadic<>` quite like we do with the inline
assembly (and TableGen code).

So we add a custom parser/printer for both (mainly stolen from TableGen
implementation, replacing parseTypeList with parseCommaSeparatedList and
a similar lambda. This takes care of the GCC code generation crash.

Fixes #45