Custom parser and printer for xsmm dispatch
Implement a custom parser and printer for xsmm dispatch ops. Currently
only for matmul and brgemm but all the dispatch ops will follow. The
custom parser and printer is necessary to preserver the pretty printing
of the enum, which otherwise would get printed as integer. I did not
find a way to get the same result using a custom assembly form.