Retire `use-extract-metadata` option
Aligned the IREE and TPP runtime to use the same calling convention.
Modify the TPP runtime to pass the aligned pointer and the offset
instead of the entire memref descriptors. This is still not ideal
because we still do some offset computation in the runtime and could do
it simply at compile time. But at least now we have a unified Xsmm to
func lowering for IREE and TPP.

Co-authored-by: Adam Siemieniuk <adam.siemieniuk@intel.com>