Restrict tensor pack folding to few cases (#521)
Previously, the following IR:

```
%cst = arith.constant 0 : i64
%empty = tensor.empty() : tensor<3xi64>
%fill = linalg.fill ins(%cts : i64) outs(%empty : tensor<3xi64>) -> tensor<3xi64>
pack(%fill)

```
was folded into:

```
%cts = arith.constant dense<0> : i64
```

This is too aggressive because it may lead to a situation where the
writable output of a given operation is replaced with a read-only
constant. Similarly,

```
func.func @main(%arg0: tensor<3x3xf32>) -> tensor<1x1x3x3xf32>) {
   %cst = arith.constant 0 : i64
   %pack (%cst) into %arg0
   return %pack
}
```
is not allowed to fold as the user intention is unclear. Is `%arg0` the
final destination?