Retire VNNI dialect
Retire VNNI dialect and VNNI TPP operations.

History: We introduced the VNNI dialect before TPP on tensor. Tensor
semantics was needed as packing happens at this abstraction. Now that we
have TPP on tensor, we no longer need the VNNI dialect.

Annoyance: We use an explicit VNNI layout (i.e., the layout is reflected
on the tensor/memef shape); this complicates the verification for the
ops and the lowering to XSMM and Loops. We will move VNNI after
bufferization and this will solve the problem.