Disable transforms for default pipeline
We currently don't support transforms together with the default
pipeline, so we shouldn't try to run it regardless. Once we start
integrating transforms with the rest of the pipeline, we must have tests
that show the integration as we activate more and more transforms.

For now, transforms need to be run manually via tpp-opt with the right
flags.

Fixes #298