Adds fused linalg addf + maxf detection
This adds a generic detection for multiple chained ops in a linalg body
to allow for fused element-wise addf + maxf that are generated by IREE.

It introduces a generic mechanism that duplicated a bit of the one-op
matcher but that will eventually replace no-op and one-op matcher with a
generic multi-op (zero or more) detector.

This can later be used for matmul (mulf + addf), groups (bcast + gemm +
bias + relu) and equations.

Fixes #571