Remove heap-to-stack pass (#578)
Removes 'heap-to-stack' pass.

The pass was originally introduced as a hack to address performance
issues in some of the benchmarks. However, it is a hacky approach and
generally leads to huge stack usage. Thus, its benefits are limited.
With the improved code generation and better benchmark kernels, this
pass is no longer needed.