Add predicate to match operand rank
Commit 0ff739123309ac8e1ab68f6dbb3dadce9f1ee27f restricted tpp operation
to 2d output, but the matchers from linalg to tpp were not constrained,
leading to verification errors. Introduce a matching predicate that
allows constraining the rank of the operand, specifically not the
matcher will filter only inputs that have rank 0 (scalar), 1 or 2. In
contrast, for output, only rank 2 is allowed.