[TPP] Fix in-place bufferization (#588)
To bufferize in place a binary operation where the lhs and rhs are the
same operands, we were ignoring a conflict between the operand, but we
can ignore this conflict only if the operand is dead after the current
operation. If we do not ignore the conflict bufferization sees a
read/write effect on the same operand and introduces copies.