Improve `packVNNIMatmulOp` (#595)
`packVNNIMatmulOp` was using hard-coded checks to pack a generic matmul
operation. The "if" branch was VNNI-blocking an already blocked matmul,
while the "else" branch was a "tiled" version. With a closer look at the
"else" branch, it turned out that the generic is a simple brgemm
operation, as a tile-blocked matmul is a brgemm. Thus remove the "else"
branch and use a de-generalization pattern to get a named op. At this
point, `PackVNNIBrgemm` can do the job of the "else" branch.