clang-format check
Adds a clang-format script and CMake target to allow developers to
monitor the conformance of their code to code formatting.

Also adding the `clang-format` target to `check-all` so that we're
always told when we're failing the formatting rules.

The `clang-format.sh` script also has a `-i` mode that actualy changes
the formatting in-place. This is meant for developers that are told
they're breaking the rules, not for CI to magically change the
formatting.

The usage is:
```
  editor> ... development ...
  build> ninja check-all
         ...
         2 files contain formatting errors
  src> ./script/ci/clang-format.sh -i
       ...
  build> ninja check-all
         ...
         All tests pass!
```

Using `clang-check-16` as it's the latest and the one we use in CI. It's
also easy to install from Debian based disros and Arch Linux. RedHat
distros can get via git-clang-format.

Fixes #583