tpp-run - replace warmup call with extra iteration (#630)
Replaces the initial warmup call with an extra timer loop iteration.
This approach simplifies kernel wrapper logic and ensures no extra
allocs or copies are required to handle the output of the separate
warmup call.

Output printing is disabled for runs with 'n > 1' as the output buffer
gets reused on each timer without any guarantees to produce stable
results on every iteration. And, in general, there is no need to get
both timing results and print output in a single command. This also
simplifies generated wrapper code.