Retire the rewrite to BRGEMM pass (#666)
Fusion now is general enough and the tiling strategy used in tile and
fuse already exposes BRGEMMs. Thus, `rewrite-to-brgemm` becomes a simple
trivial rewrite. Adjust fusion to tile if the working group consists of
a single operation, we need to do this as we benchmarks and map 
single operations (for single op we tile only). 
Previously, fusion was assuming a working group
of at least two operations.

Fixes: https://github.com/plaidml/tpp-mlir/issues/392