Move `createRewriteBatchMatmulToMatmulPass`  (#663)
There is no reason to keep `createRewriteBatchMatmulToMatmulPass` as
pass in `createTppConversionPass`. It is not related to tpp, it is a
linalg to linalg pass.

Rename:
- linalg-to-loops.mlir
- linalg-tensor.mlir

More descriptive names.

Regression:

In `pass-rewrite-batch-matmul-to-matmul.mlir` the fully dynamic case is
not mapped anymore to a linalg.matmul. ForAll tiling insert an affine
min that prevents rank reduction. Unlcear what this is the case.