Enable linalg-to-xsmm in tpp-run (#679)
Enable `linalg-to-xsmm` in tpp.run and fix the mapping issue for binary and unary. For unary and binary operations, we need to use the captured operands and not the operands of the linalg generic. For example,  a relu in linalg can be expressed using a single operand, while in XSMM, we require two operands for the unary (input and output).