Linalg to XSMM: BrgemmOp conversion (#678)
Allows mapping generic contractions to BRGEMM. The idea is to find a
contraction operation with at least two parallel dimensions (m and n)
and, at most, two reduction dimensions (k and k'). The m dimension
should appear on the LHS and the OUT operand but not the RHS, while the
n dimension should appear on the RHS and the OUT operand but not the
LHS. k and k' should appear only on the OUT. m, n, and k are selected as
minor dimensions for the GEMM and must have an innermost stride of 1,
while the k' is chosen as batch reduction dimension.