Run fusion for a fixed-point number of iterations (#697)
When running more complex benchmarks (i.e., MHA), fusion opportunities
can be exposed by running the pass multiple times. Currently, the pass
runs three times as this is needed for MHA; we may want more
sophisticated heuristics in the future. Retire also the option
`start-from-last-fusable-consumer` as this is always profitable. Drop
`tile-and-fuse-with-cse.mlir` as the allocation in the loops is needed
because it iterates over a tile of 32x32.