Fix scf.forall to scf.parallel conversion (#738)
Upstream converts `scf.forall` to `scf.parallel`, which is, in turn,
lowered to an `scf.for` and then to structured control flow in
`convert-scf-to-cf`. This lowering is atomic, thus we need to
materialize `scf.parallel` before `convert-scf-to-cf` if we want to map
to `omp.parallel`. The ideal solution would be to decouple the lowering
from `scf.forall` to `scf.parallel` from upstream: `convert-scf-to-cf`.

fixes: #734