Refactor all kernels into a single generator
Avoid duplicating the logic across `matmul`, `fc` and `mlp` into a
single generator that enables/disables features based on command line
flags.

Also enables multi-layer for `fc` and `matmul` that previously wasn't
possible.

These are the new choices:
 * Kernel = `layer` (operands are arguments), `model` (constants)
 * Enable/disable bias/relu/softmax

The matching is:
 * `matmul` = kernel=layer
 * `fc` = kernel=layer + bias + relu
 * `mlp` = kernel=model + bias + relu

There should be a softmax at the end of MLP, but we're not doing it yet.

All tests pass. All benchmarks kept their performance profiles.

Changes in tests are due to real (and necessary) output changes.