Retire populateTppDeGeneralizationPatterns (#746)
We are (slowly) retiring tpp and directly jump to xsmm from linalg. Do
not raise back to tpp
after fusion to avoid having tpp ops in the linalg to xsmm lowerings.

I decided to remove: `mlp-all` as it is a poorly written test, I can add
more vnni tests later on.