Add debug scripts for compiler passes
This is a simple set of scripts that run mlir-gen/tpp-opt, dumps the IR
between passes, splits the output for each pass into files, looks at
each one when the IR changes and runs the diff tool of your choice to
show these differences.

For now, mlir-gen is only emitting the standard odd-shaped MLP layer
over the standard TPP passes, but future changes to the scripts can
add arguments to each tool.