Allow fusion of fill with multiple users (#767)
If an operation has multiple users, not all in the same fusion domain,
fusion does not take the operation into account. Allowing operations
with multiple users in various fusion domains introduces recomputation.
Introduce an exception for FillOp. This allows us to bring fillOps
operation in the fusion domain even after CSE and later mapping the
operation to XSMM calls. Without this change we end up with scalar fills
or copy.