XSMM Bump to HEAD of main (#775)
This bumps XSMM and XSMM-DNN to HEAD of main on both repos. We still get
the warnings (#707) but at least now we're on a known commit.

Once #707 and https://github.com/libxsmm/libxsmm-dnn/issues/21 get
fixed, we can bump again and enforce warnings as errors, so that we
don't get to this place again.