Require expected rank in `isInVnniLayout` (#809)
Until we have a better way to express the VNNI layout (see: #563), it is
up to the caller to specify the expected rank in the VNNI layout as the
rank depends on the operations we are dealing with. The current check
for rank < 3 is not correct; this is still ugly, but at least more
explicit, as there are not magic assumptions on rank.