[Perf] Move timer out of the loop (#807)
This is a big change to the perf dialect: we're moving the timer
start/stop to outside of the loop and just dividing it by the number of
iterations, instead of calculating mean and stdev for each timer
start/stop inside the loop.

Benchmarks on the cluster have shown identical results to all previous
machines we tested on, so there's no actual different on the run time.

But the code is significantly better:
 * No need to implement mean/stdev logic
 * No need to allocate results tensor, insert/extract
 * Higher precision with low-precision timers
 * Significantly simpler bench.perf logic and lowering

The main changes are:
 * Perf dialect to remove unused ops and simplify bench
 * tpp-run to produce the right warmup, bench loop
 * benchmark driver to not print stdev on results

Some related changes:
 * Reinforced tpp-run to not print void functions
 * Warmup loop of 1% num_iter between 2 and 50

---------

Co-authored-by: Adam Siemieniuk <adam.siemieniuk@intel.com>