[CI] Persisting some changes done during benchmarking session (#814)
Some of the changes we did this week should be persisted, as they were
good changes:
* Reorder some benchmarks in the result to make it easier to copy&paste
into a spreadsheet. No change in the benchmarks themselves.
* Add PyTorch OpenMP tests in the same way as tpp-mlir and libxsmm-dnn.
This gives us a packing scalability measurement.
* Split base / openmp into two different jobs to increase scalability of
benchmark CI jobs (run in parallel). We just need to be careful with the
machines we don't have too many, so for now it's only enabled by default
on SPR.
* Disabling unnecessary build of OneDNN in benchmarks until we actually
benchmark OneDNN calls.