Change `mlir-gen` to create bias with broadcast (#821)
This commit creates models with bias constant 1D and broadcasts it into
the bias add after matmul on MLP models. This is the most common case
upstream and is also the only case where we support fused BRGEMM in
libxsmm.

I'm explicitly avoiding an extra option because this part of the code is
already too complicated to fiddle with and we'll soon be using real IR
(from PyTorch etc) to test the passes.