[NFC] Rename model/layer to inference/training in mlir-gen (#828)
This doesn't change anything, just renames the argument to make it
clearer what it means.

What we used to call 'layer', because we had one function per layer and
thus passed all arguments (including weights and biases) as arguments,
is now called 'inference', because we only use arguments for weights and
biases when they're actually not constant.

Therefore, the previous 'model' is now called 'inference', because we
initialize the weights and biases as constants, as if they've been
pre-trained and cannot be changed.

Since we don't do the one-function-per-layer anymore, these are the only
two use cases for such kernel types, so we make the argument clear to
avoid confusion when creating tests.

Right now, we're focusing on inference code, but it's good to have tests
and benchmarks for training mode anyway.