Update fusion after llvm bump
Previously, only building blocks were exposed, and fusion users were
supposed to assemble them. Most heavy lifting is shifted into
`tileConsumerAndFuseProducerGreedilyUsingSCFForOp,` which tiles a
consumer and fuses it with producers. Note that the method relies on a
callback to decide what to fuse; what to fuse is the tricky part, which
for us is computed in `collectFusableProducers`.