[NFC] Separate local dialect lowering from low-level parallelization (#916)
The 2D parallelization and AMX tile config stuff were bundled into the
local dialect lowering passes because that's "an XSMM thing". But that's
changing, as we start working on multi-level loop parallelization, this
will raise to tensor eventually and we need to separate the transform
from the dialect.

This PR however, does none of that. It just separates the "bundles" to
de-couple dialect lowering from the loop passes. Should not change a
iota on the output.

Also removing a duplicated call to lower the `perf` dialect in the main
pipeline.

---------

Co-authored-by: Adam Siemieniuk <adam.siemieniuk@intel.com>